legacy system ported to python 3
