+++ legacy/src/Makefile
@@ -1,11 +1,11 @@
 .PHONY: init run test
 
 init:
-	PYTHONPATH=backend python2 -c "from app import init_db; init_db()"
+	PYTHONPATH=backend python3 -c "from database import init_db; init_db()"
 
 run:
-	python2 ./backend/app.py
+	PYTHONPATH=backend python3 -c "from database import check_db_exists; check_db_exists()" || make init
+	PYTHONPATH=backend python3 backend/app.py
 
-test: 
-	PYTHONPATH=backend python2 ./backend/app_tests.py
-
+test:
+	PYTHONPATH=backend python3 backend/app_tests.py
