+++ legacy/src/backend/config.py
@@ -1,0 +1,4 @@
+DATABASE_PATH = './whoknows.db'
+PER_PAGE = 30
+DEBUG = False
+SECRET_KEY = 'development key'
