+++ legacy/src/backend/requirements.txt
@@ -1,4 +1,4 @@
-setuptools==44.1.1
-Werkzeug==0.6.1
-Jinja2==2.4
-Flask==0.5
+setuptools==74.0.0
+Werkzeug==3.0.4
+Jinja2==3.1.4
+Flask==3.0.3
